refactor: new interface FakeClock
