+++ pkg/time/fakeclock.go
@@ -1,0 +1,23 @@
+// Copyright 2021 Chaos Mesh Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package time
+
+// FakeClock could modify the time of certain process.
+// TODO: rename this interface, it brings confusing with .c files under pkg/time/fakeclock.
+type FakeClock interface {
+	Inject(pid int) error
+	Recover(pid int) error
+}
